pretty much done, might need some tidying up
