+++ baseapp/baseapp_test.go
@@ -1,396 +1,46 @@
 package baseapp
 
 import (
-	"bytes"
-	"encoding/binary"
-	"encoding/json"
-	"fmt"
-	"math/rand"
-	"os"
 	"testing"
-	"time"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/log"
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
 	dbm "github.com/tendermint/tm-db"
 
-	"github.com/cosmos/cosmos-sdk/codec"
-	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	pruningtypes "github.com/cosmos/cosmos-sdk/pruning/types"
-	"github.com/cosmos/cosmos-sdk/snapshots"
-	snapshottypes "github.com/cosmos/cosmos-sdk/snapshots/types"
-	"github.com/cosmos/cosmos-sdk/store/rootmulti"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
-	"github.com/cosmos/cosmos-sdk/testutil"
-	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 )
 
 var (
 	capKey1 = sdk.NewKVStoreKey("key1")
 	capKey2 = sdk.NewKVStoreKey("key2")
-
-	// testTxPriority is the CheckTx priority that we set in the test
-	// antehandler.
-	testTxPriority = int64(42)
 )
 
-type paramStore struct {
-	db *dbm.MemDB
+func TestSetMinGasPrices(t *testing.T) {
+	minGasPrices := sdk.DecCoins{sdk.NewInt64DecCoin("stake", 5000)}
+	app := setupBaseApp(t, SetMinGasPrices(minGasPrices.String()))
+	require.Equal(t, minGasPrices, app.minGasPrices)
 }
 
-type setupConfig struct {
-	blocks             uint64
-	blockTxs           int
-	snapshotInterval   uint64
-	snapshotKeepRecent uint32
-	pruningOpts        pruningtypes.PruningOptions
-}
-
-func (ps *paramStore) Set(_ sdk.Context, key []byte, value interface{}) {
-	bz, err := json.Marshal(value)
-	if err != nil {
-		panic(err)
-	}
-
-	ps.db.Set(key, bz)
-}
-
-func (ps *paramStore) Has(_ sdk.Context, key []byte) bool {
-	ok, err := ps.db.Has(key)
-	if err != nil {
-		panic(err)
-	}
-
-	return ok
-}
-
-func (ps *paramStore) Get(_ sdk.Context, key []byte, ptr interface{}) {
-	bz, err := ps.db.Get(key)
-	if err != nil {
-		panic(err)
-	}
-
-	if len(bz) == 0 {
-		return
-	}
-
-	if err := json.Unmarshal(bz, ptr); err != nil {
-		panic(err)
-	}
-}
-
-func defaultLogger() log.Logger {
-	return log.NewTMLogger(log.NewSyncWriter(os.Stdout)).With("module", "sdk/app")
-}
-
-func newBaseApp(name string, options ...func(*BaseApp)) *BaseApp {
-	logger := defaultLogger()
-	db := dbm.NewMemDB()
-	codec := codec.NewLegacyAmino()
-	registerTestCodec(codec)
-	return NewBaseApp(name, logger, db, testTxDecoder(codec), options...)
-}
-
-func registerTestCodec(cdc *codec.LegacyAmino) {
-	// register Tx, Msg
-	sdk.RegisterLegacyAminoCodec(cdc)
-
-	// register test types
-	// cdc.RegisterConcrete(&txTest{}, "cosmos-sdk/baseapp/txTest", nil)
-	// cdc.RegisterConcrete(&msgCounter{}, "cosmos-sdk/baseapp/msgCounter", nil)
-	// cdc.RegisterConcrete(&msgCounter2{}, "cosmos-sdk/baseapp/msgCounter2", nil)
-	// cdc.RegisterConcrete(&msgKeyValue{}, "cosmos-sdk/baseapp/msgKeyValue", nil)
-	// cdc.RegisterConcrete(&msgNoRoute{}, "cosmos-sdk/baseapp/msgNoRoute", nil)
-}
-
-// aminoTxEncoder creates a amino TxEncoder for testing purposes.
-func aminoTxEncoder() sdk.TxEncoder {
-	cdc := codec.NewLegacyAmino()
-	registerTestCodec(cdc)
-
-	return legacytx.StdTxConfig{Cdc: cdc}.TxEncoder()
-}
-
-// simple one store baseapp
-func setupBaseApp(t *testing.T, options ...func(*BaseApp)) *BaseApp {
-	app := newBaseApp(t.Name(), options...)
-	require.Equal(t, t.Name(), app.Name())
-
-	app.MountStores(capKey1, capKey2)
-	app.SetParamStore(&paramStore{db: dbm.NewMemDB()})
-
-	// stores are mounted
-	err := app.LoadLatestVersion()
-	require.Nil(t, err)
-	return app
-}
-
-// simple one store baseapp with data and snapshots. Each tx is 1 MB in size (uncompressed).
-func setupBaseAppWithSnapshots(t *testing.T, config *setupConfig) (*BaseApp, error) {
-	codec := codec.NewLegacyAmino()
-	registerTestCodec(codec)
-	routerOpt := func(bapp *BaseApp) {
-		// bapp.Router().AddRoute(sdk.NewRoute(routeMsgKeyValue, func(ctx sdk.Context, msg sdk.Msg) (*sdk.Result, error) {
-		// 	kv := msg.(*msgKeyValue)
-		// 	bapp.cms.GetCommitKVStore(capKey2).Set(kv.Key, kv.Value)
-		// 	return &sdk.Result{}, nil
-		// }))
-	}
-
-	snapshotTimeout := 1 * time.Minute
-	snapshotStore, err := snapshots.NewStore(dbm.NewMemDB(), testutil.GetTempDir(t))
-	require.NoError(t, err)
-
-	app := setupBaseApp(t, routerOpt, SetSnapshot(snapshotStore, snapshottypes.NewSnapshotOptions(config.snapshotInterval, uint32(config.snapshotKeepRecent))), SetPruning(config.pruningOpts))
-
-	app.InitChain(abci.RequestInitChain{})
-
-	r := rand.New(rand.NewSource(3920758213583))
-	keyCounter := 0
-	for height := int64(1); height <= int64(config.blocks); height++ {
-		app.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: height}})
-		for txNum := 0; txNum < config.blockTxs; txNum++ {
-			tx := txTest{Msgs: []sdk.Msg{}}
-			for msgNum := 0; msgNum < 100; msgNum++ {
-				key := []byte(fmt.Sprintf("%v", keyCounter))
-				value := make([]byte, 10000)
-				_, err := r.Read(value)
-				require.NoError(t, err)
-				tx.Msgs = append(tx.Msgs, msgKeyValue{Key: key, Value: value})
-				keyCounter++
-			}
-			txBytes, err := codec.Marshal(tx)
-			require.NoError(t, err)
-			resp := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-			require.True(t, resp.IsOK(), "%v", resp.String())
-		}
-		app.EndBlock(abci.RequestEndBlock{Height: height})
-		app.Commit()
-
-		// Wait for snapshot to be taken, since it happens asynchronously.
-		if config.snapshotInterval > 0 && uint64(height)%config.snapshotInterval == 0 {
-			start := time.Now()
-			for {
-				if time.Since(start) > snapshotTimeout {
-					t.Errorf("timed out waiting for snapshot after %v", snapshotTimeout)
-				}
-				snapshot, err := snapshotStore.Get(uint64(height), snapshottypes.CurrentFormat)
-				require.NoError(t, err)
-				if snapshot != nil {
-					break
-				}
-				time.Sleep(100 * time.Millisecond)
-			}
-		}
-	}
-
-	return app, nil
-}
-
-func TestMountStores(t *testing.T) {
+func TestGetMaximumBlockGas(t *testing.T) {
 	app := setupBaseApp(t)
+	app.InitChain(abci.RequestInitChain{})
+	ctx := app.NewContext(true, tmproto.Header{})
 
-	// check both stores
-	store1 := app.cms.GetCommitKVStore(capKey1)
-	require.NotNil(t, store1)
-	store2 := app.cms.GetCommitKVStore(capKey2)
-	require.NotNil(t, store2)
-}
+	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: 0}})
+	require.Equal(t, uint64(0), app.getMaximumBlockGas(ctx))
 
-// Test that we can make commits and then reload old versions.
-// Test that LoadLatestVersion actually does.
-func TestLoadVersion(t *testing.T) {
-	logger := defaultLogger()
-	pruningOpt := SetPruning(pruningtypes.NewPruningOptions(pruningtypes.PruningNothing))
-	db := dbm.NewMemDB()
-	name := t.Name()
-	app := NewBaseApp(name, logger, db, nil, pruningOpt)
+	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: -1}})
+	require.Equal(t, uint64(0), app.getMaximumBlockGas(ctx))
 
-	// make a cap key and mount the store
-	err := app.LoadLatestVersion() // needed to make stores non-nil
-	require.Nil(t, err)
+	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: 5000000}})
+	require.Equal(t, uint64(5000000), app.getMaximumBlockGas(ctx))
 
-	emptyCommitID := storetypes.CommitID{}
-
-	// fresh store has zero/empty last commit
-	lastHeight := app.LastBlockHeight()
-	lastID := app.LastCommitID()
-	require.Equal(t, int64(0), lastHeight)
-	require.Equal(t, emptyCommitID, lastID)
-
-	// execute a block, collect commit ID
-	header := tmproto.Header{Height: 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	res := app.Commit()
-	commitID1 := storetypes.CommitID{Version: 1, Hash: res.Data}
-
-	// execute a block, collect commit ID
-	header = tmproto.Header{Height: 2}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	res = app.Commit()
-	commitID2 := storetypes.CommitID{Version: 2, Hash: res.Data}
-
-	// reload with LoadLatestVersion
-	app = NewBaseApp(name, logger, db, nil, pruningOpt)
-	app.MountStores()
-	err = app.LoadLatestVersion()
-	require.Nil(t, err)
-	testLoadVersionHelper(t, app, int64(2), commitID2)
-
-	// reload with LoadVersion, see if you can commit the same block and get
-	// the same result
-	app = NewBaseApp(name, logger, db, nil, pruningOpt)
-	err = app.LoadVersion(1)
-	require.Nil(t, err)
-	testLoadVersionHelper(t, app, int64(1), commitID1)
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	app.Commit()
-	testLoadVersionHelper(t, app, int64(2), commitID2)
-}
-
-func useDefaultLoader(app *BaseApp) {
-	app.SetStoreLoader(DefaultStoreLoader)
-}
-
-func initStore(t *testing.T, db dbm.DB, storeKey string, k, v []byte) {
-	rs := rootmulti.NewStore(db, log.NewNopLogger())
-	rs.SetPruning(pruningtypes.NewPruningOptions(pruningtypes.PruningNothing))
-	key := sdk.NewKVStoreKey(storeKey)
-	rs.MountStoreWithDB(key, storetypes.StoreTypeIAVL, nil)
-	err := rs.LoadLatestVersion()
-	require.Nil(t, err)
-	require.Equal(t, int64(0), rs.LastCommitID().Version)
-
-	// write some data in substore
-	kv, _ := rs.GetStore(key).(storetypes.KVStore)
-	require.NotNil(t, kv)
-	kv.Set(k, v)
-	commitID := rs.Commit()
-	require.Equal(t, int64(1), commitID.Version)
-}
-
-func checkStore(t *testing.T, db dbm.DB, ver int64, storeKey string, k, v []byte) {
-	rs := rootmulti.NewStore(db, log.NewNopLogger())
-	rs.SetPruning(pruningtypes.NewPruningOptions(pruningtypes.PruningDefault))
-	key := sdk.NewKVStoreKey(storeKey)
-	rs.MountStoreWithDB(key, storetypes.StoreTypeIAVL, nil)
-	err := rs.LoadLatestVersion()
-	require.Nil(t, err)
-	require.Equal(t, ver, rs.LastCommitID().Version)
-
-	// query data in substore
-	kv, _ := rs.GetStore(key).(storetypes.KVStore)
-	require.NotNil(t, kv)
-	require.Equal(t, v, kv.Get(k))
-}
-
-// Test that we can make commits and then reload old versions.
-// Test that LoadLatestVersion actually does.
-func TestSetLoader(t *testing.T) {
-	cases := map[string]struct {
-		setLoader    func(*BaseApp)
-		origStoreKey string
-		loadStoreKey string
-	}{
-		"don't set loader": {
-			origStoreKey: "foo",
-			loadStoreKey: "foo",
-		},
-		"default loader": {
-			setLoader:    useDefaultLoader,
-			origStoreKey: "foo",
-			loadStoreKey: "foo",
-		},
-	}
-
-	k := []byte("key")
-	v := []byte("value")
-
-	for name, tc := range cases {
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			// prepare a db with some data
-			db := dbm.NewMemDB()
-			initStore(t, db, tc.origStoreKey, k, v)
-
-			// load the app with the existing db
-			opts := []func(*BaseApp){SetPruning(pruningtypes.NewPruningOptions(pruningtypes.PruningNothing))}
-			if tc.setLoader != nil {
-				opts = append(opts, tc.setLoader)
-			}
-			app := NewBaseApp(t.Name(), defaultLogger(), db, nil, opts...)
-			app.MountStores(sdk.NewKVStoreKey(tc.loadStoreKey))
-			err := app.LoadLatestVersion()
-			require.Nil(t, err)
-
-			// "execute" one block
-			app.BeginBlock(abci.RequestBeginBlock{Header: tmproto.Header{Height: 2}})
-			res := app.Commit()
-			require.NotNil(t, res.Data)
-
-			// check db is properly updated
-			checkStore(t, db, 2, tc.loadStoreKey, k, v)
-			checkStore(t, db, 2, tc.loadStoreKey, []byte("foo"), nil)
-		})
-	}
-}
-
-func TestVersionSetterGetter(t *testing.T) {
-	logger := defaultLogger()
-	pruningOpt := SetPruning(pruningtypes.NewPruningOptions(pruningtypes.PruningDefault))
-	db := dbm.NewMemDB()
-	name := t.Name()
-	app := NewBaseApp(name, logger, db, nil, pruningOpt)
-
-	require.Equal(t, "", app.Version())
-	res := app.Query(abci.RequestQuery{Path: "app/version"})
-	require.True(t, res.IsOK())
-	require.Equal(t, "", string(res.Value))
-
-	versionString := "1.0.0"
-	app.SetVersion(versionString)
-	require.Equal(t, versionString, app.Version())
-	res = app.Query(abci.RequestQuery{Path: "app/version"})
-	require.True(t, res.IsOK())
-	require.Equal(t, versionString, string(res.Value))
-}
-
-func TestLoadVersionInvalid(t *testing.T) {
-	logger := log.NewNopLogger()
-	pruningOpt := SetPruning(pruningtypes.NewPruningOptions(pruningtypes.PruningNothing))
-	db := dbm.NewMemDB()
-	name := t.Name()
-	app := NewBaseApp(name, logger, db, nil, pruningOpt)
-
-	err := app.LoadLatestVersion()
-	require.Nil(t, err)
-
-	// require error when loading an invalid version
-	err = app.LoadVersion(-1)
-	require.Error(t, err)
-
-	header := tmproto.Header{Height: 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	res := app.Commit()
-	commitID1 := storetypes.CommitID{Version: 1, Hash: res.Data}
-
-	// create a new app with the stores mounted under the same cap key
-	app = NewBaseApp(name, logger, db, nil, pruningOpt)
-
-	// require we can load the latest version
-	err = app.LoadVersion(1)
-	require.Nil(t, err)
-	testLoadVersionHelper(t, app, int64(1), commitID1)
-
-	// require error when loading an invalid version
-	err = app.LoadVersion(2)
-	require.Error(t, err)
+	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: -5000000}})
+	require.Panics(t, func() { app.getMaximumBlockGas(ctx) })
 }
 
 func TestLoadVersionPruning(t *testing.T) {
@@ -445,871 +95,25 @@ func TestLoadVersionPruning(t *testing.T) {
 	testLoadVersionHelper(t, app, int64(7), lastCommitID)
 }
 
+// simple one store baseapp
+func setupBaseApp(t *testing.T, options ...func(*BaseApp)) *BaseApp {
+	logger := defaultLogger()
+	db := dbm.NewMemDB()
+	app := NewBaseApp(t.Name(), logger, db, nil, options...)
+	require.Equal(t, t.Name(), app.Name())
+
+	app.MountStores(capKey1, capKey2)
+	app.SetParamStore(&paramStore{db: dbm.NewMemDB()})
+
+	// stores are mounted
+	err := app.LoadLatestVersion()
+	require.Nil(t, err)
+	return app
+}
+
 func testLoadVersionHelper(t *testing.T, app *BaseApp, expectedHeight int64, expectedID storetypes.CommitID) {
 	lastHeight := app.LastBlockHeight()
 	lastID := app.LastCommitID()
 	require.Equal(t, expectedHeight, lastHeight)
 	require.Equal(t, expectedID, lastID)
-}
-
-func TestOptionFunction(t *testing.T) {
-	logger := defaultLogger()
-	db := dbm.NewMemDB()
-	bap := NewBaseApp("starting name", logger, db, nil, testChangeNameHelper("new name"))
-	require.Equal(t, bap.name, "new name", "BaseApp should have had name changed via option function")
-}
-
-func testChangeNameHelper(name string) func(*BaseApp) {
-	return func(bap *BaseApp) {
-		bap.name = name
-	}
-}
-
-// Test that Info returns the latest committed state.
-func TestInfo(t *testing.T) {
-	app := newBaseApp(t.Name())
-
-	// ----- test an empty response -------
-	reqInfo := abci.RequestInfo{}
-	res := app.Info(reqInfo)
-
-	// should be empty
-	assert.Equal(t, "", res.Version)
-	assert.Equal(t, t.Name(), res.GetData())
-	assert.Equal(t, int64(0), res.LastBlockHeight)
-	require.Equal(t, []uint8(nil), res.LastBlockAppHash)
-	require.Equal(t, app.AppVersion(), res.AppVersion)
-	// ----- test a proper response -------
-	// TODO
-}
-
-func TestBaseAppOptionSeal(t *testing.T) {
-	app := setupBaseApp(t)
-
-	require.Panics(t, func() {
-		app.SetName("")
-	})
-	require.Panics(t, func() {
-		app.SetVersion("")
-	})
-	require.Panics(t, func() {
-		app.SetDB(nil)
-	})
-	require.Panics(t, func() {
-		app.SetCMS(nil)
-	})
-	require.Panics(t, func() {
-		app.SetInitChainer(nil)
-	})
-	require.Panics(t, func() {
-		app.SetBeginBlocker(nil)
-	})
-	require.Panics(t, func() {
-		app.SetEndBlocker(nil)
-	})
-	require.Panics(t, func() {
-		app.SetAnteHandler(nil)
-	})
-	require.Panics(t, func() {
-		app.SetAddrPeerFilter(nil)
-	})
-	require.Panics(t, func() {
-		app.SetIDPeerFilter(nil)
-	})
-	require.Panics(t, func() {
-		app.SetFauxMerkleMode()
-	})
-}
-
-func TestSetMinGasPrices(t *testing.T) {
-	minGasPrices := sdk.DecCoins{sdk.NewInt64DecCoin("stake", 5000)}
-	app := newBaseApp(t.Name(), SetMinGasPrices(minGasPrices.String()))
-	require.Equal(t, minGasPrices, app.minGasPrices)
-}
-
-func TestInitChainer(t *testing.T) {
-	name := t.Name()
-	// keep the db and logger ourselves so
-	// we can reload the same  app later
-	db := dbm.NewMemDB()
-	logger := defaultLogger()
-	app := NewBaseApp(name, logger, db, nil)
-	capKey := sdk.NewKVStoreKey("main")
-	capKey2 := sdk.NewKVStoreKey("key2")
-	app.MountStores(capKey, capKey2)
-
-	// set a value in the store on init chain
-	key, value := []byte("hello"), []byte("goodbye")
-	var initChainer sdk.InitChainer = func(ctx sdk.Context, req abci.RequestInitChain) abci.ResponseInitChain {
-		store := ctx.KVStore(capKey)
-		store.Set(key, value)
-		return abci.ResponseInitChain{}
-	}
-
-	query := abci.RequestQuery{
-		Path: "/store/main/key",
-		Data: key,
-	}
-
-	// initChainer is nil - nothing happens
-	app.InitChain(abci.RequestInitChain{})
-	res := app.Query(query)
-	require.Equal(t, 0, len(res.Value))
-
-	// set initChainer and try again - should see the value
-	app.SetInitChainer(initChainer)
-
-	// stores are mounted and private members are set - sealing baseapp
-	err := app.LoadLatestVersion() // needed to make stores non-nil
-	require.Nil(t, err)
-	require.Equal(t, int64(0), app.LastBlockHeight())
-
-	initChainRes := app.InitChain(abci.RequestInitChain{AppStateBytes: []byte("{}"), ChainId: "test-chain-id"}) // must have valid JSON genesis file, even if empty
-
-	// The AppHash returned by a new chain is the sha256 hash of "".
-	// $ echo -n '' | sha256sum
-	// e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
-	require.Equal(
-		t,
-		[]byte{0xe3, 0xb0, 0xc4, 0x42, 0x98, 0xfc, 0x1c, 0x14, 0x9a, 0xfb, 0xf4, 0xc8, 0x99, 0x6f, 0xb9, 0x24, 0x27, 0xae, 0x41, 0xe4, 0x64, 0x9b, 0x93, 0x4c, 0xa4, 0x95, 0x99, 0x1b, 0x78, 0x52, 0xb8, 0x55},
-		initChainRes.AppHash,
-	)
-
-	// assert that chainID is set correctly in InitChain
-	chainID := app.deliverState.ctx.ChainID()
-	require.Equal(t, "test-chain-id", chainID, "ChainID in deliverState not set correctly in InitChain")
-
-	chainID = app.checkState.ctx.ChainID()
-	require.Equal(t, "test-chain-id", chainID, "ChainID in checkState not set correctly in InitChain")
-
-	app.Commit()
-	res = app.Query(query)
-	require.Equal(t, int64(1), app.LastBlockHeight())
-	require.Equal(t, value, res.Value)
-
-	// reload app
-	app = NewBaseApp(name, logger, db, nil)
-	app.SetInitChainer(initChainer)
-	app.MountStores(capKey, capKey2)
-	err = app.LoadLatestVersion() // needed to make stores non-nil
-	require.Nil(t, err)
-	require.Equal(t, int64(1), app.LastBlockHeight())
-
-	// ensure we can still query after reloading
-	res = app.Query(query)
-	require.Equal(t, value, res.Value)
-
-	// commit and ensure we can still query
-	header := tmproto.Header{Height: app.LastBlockHeight() + 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	app.Commit()
-
-	res = app.Query(query)
-	require.Equal(t, value, res.Value)
-}
-
-func TestInitChain_WithInitialHeight(t *testing.T) {
-	name := t.Name()
-	db := dbm.NewMemDB()
-	logger := defaultLogger()
-	app := NewBaseApp(name, logger, db, nil)
-
-	app.InitChain(
-		abci.RequestInitChain{
-			InitialHeight: 3,
-		},
-	)
-	app.Commit()
-
-	require.Equal(t, int64(3), app.LastBlockHeight())
-}
-
-func TestBeginBlock_WithInitialHeight(t *testing.T) {
-	name := t.Name()
-	db := dbm.NewMemDB()
-	logger := defaultLogger()
-	app := NewBaseApp(name, logger, db, nil)
-
-	app.InitChain(
-		abci.RequestInitChain{
-			InitialHeight: 3,
-		},
-	)
-
-	require.PanicsWithError(t, "invalid height: 4; expected: 3", func() {
-		app.BeginBlock(abci.RequestBeginBlock{
-			Header: tmproto.Header{
-				Height: 4,
-			},
-		})
-	})
-
-	app.BeginBlock(abci.RequestBeginBlock{
-		Header: tmproto.Header{
-			Height: 3,
-		},
-	})
-	app.Commit()
-
-	require.Equal(t, int64(3), app.LastBlockHeight())
-}
-
-// Simple tx with a list of Msgs.
-type txTest struct {
-	Msgs       []sdk.Msg
-	Counter    int64
-	FailOnAnte bool
-}
-
-func (tx *txTest) setFailOnAnte(fail bool) {
-	tx.FailOnAnte = fail
-}
-
-func (tx *txTest) setFailOnHandler(fail bool) {
-	for i, msg := range tx.Msgs {
-		tx.Msgs[i] = msgCounter{msg.(msgCounter).Counter, fail}
-	}
-}
-
-// Implements Tx
-func (tx txTest) GetMsgs() []sdk.Msg   { return tx.Msgs }
-func (tx txTest) ValidateBasic() error { return nil }
-
-const (
-	routeMsgCounter  = "msgCounter"
-	routeMsgCounter2 = "msgCounter2"
-	routeMsgKeyValue = "msgKeyValue"
-)
-
-// ValidateBasic() fails on negative counters.
-// Otherwise it's up to the handlers
-type msgCounter struct {
-	Counter       int64
-	FailOnHandler bool
-}
-
-// dummy implementation of proto.Message
-func (msg msgCounter) Reset()         {}
-func (msg msgCounter) String() string { return "TODO" }
-func (msg msgCounter) ProtoMessage()  {}
-
-// Implements Msg
-func (msg msgCounter) Route() string                { return routeMsgCounter }
-func (msg msgCounter) Type() string                 { return "counter1" }
-func (msg msgCounter) GetSignBytes() []byte         { return nil }
-func (msg msgCounter) GetSigners() []sdk.AccAddress { return nil }
-func (msg msgCounter) ValidateBasic() error {
-	if msg.Counter >= 0 {
-		return nil
-	}
-	return sdkerrors.Wrap(sdkerrors.ErrInvalidSequence, "counter should be a non-negative integer")
-}
-
-func newTxCounter(counter int64, msgCounters ...int64) *txTest {
-	msgs := make([]sdk.Msg, 0, len(msgCounters))
-	for _, c := range msgCounters {
-		msgs = append(msgs, msgCounter{c, false})
-	}
-
-	return &txTest{msgs, counter, false}
-}
-
-// A msg that sets a key/value pair.
-type msgKeyValue struct {
-	Key   []byte
-	Value []byte
-}
-
-func (msg msgKeyValue) Reset()                       {}
-func (msg msgKeyValue) String() string               { return "TODO" }
-func (msg msgKeyValue) ProtoMessage()                {}
-func (msg msgKeyValue) Route() string                { return routeMsgKeyValue }
-func (msg msgKeyValue) Type() string                 { return "keyValue" }
-func (msg msgKeyValue) GetSignBytes() []byte         { return nil }
-func (msg msgKeyValue) GetSigners() []sdk.AccAddress { return nil }
-func (msg msgKeyValue) ValidateBasic() error {
-	if msg.Key == nil {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "key cannot be nil")
-	}
-	if msg.Value == nil {
-		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "value cannot be nil")
-	}
-	return nil
-}
-
-// amino decode
-func testTxDecoder(cdc *codec.LegacyAmino) sdk.TxDecoder {
-	return func(txBytes []byte) (sdk.Tx, error) {
-		var tx txTest
-		if len(txBytes) == 0 {
-			return nil, sdkerrors.Wrap(sdkerrors.ErrTxDecode, "tx bytes are empty")
-		}
-
-		err := cdc.Unmarshal(txBytes, &tx)
-		if err != nil {
-			return nil, sdkerrors.ErrTxDecode
-		}
-
-		return tx, nil
-	}
-}
-
-func anteHandlerTxTest(t *testing.T, capKey storetypes.StoreKey, storeKey []byte) sdk.AnteHandler {
-	return func(ctx sdk.Context, tx sdk.Tx, simulate bool) (sdk.Context, error) {
-		store := ctx.KVStore(capKey)
-		txTest := tx.(txTest)
-
-		if txTest.FailOnAnte {
-			return ctx, sdkerrors.Wrap(sdkerrors.ErrUnauthorized, "ante handler failure")
-		}
-
-		_, err := incrementingCounter(t, store, storeKey, txTest.Counter)
-		if err != nil {
-			return ctx, err
-		}
-
-		ctx.EventManager().EmitEvents(
-			counterEvent("ante_handler", txTest.Counter),
-		)
-
-		ctx = ctx.WithPriority(testTxPriority)
-
-		return ctx, nil
-	}
-}
-
-func counterEvent(evType string, msgCount int64) sdk.Events {
-	return sdk.Events{
-		sdk.NewEvent(
-			evType,
-			sdk.NewAttribute("update_counter", fmt.Sprintf("%d", msgCount)),
-		),
-	}
-}
-
-func getIntFromStore(store sdk.KVStore, key []byte) int64 {
-	bz := store.Get(key)
-	if len(bz) == 0 {
-		return 0
-	}
-	i, err := binary.ReadVarint(bytes.NewBuffer(bz))
-	if err != nil {
-		panic(err)
-	}
-	return i
-}
-
-func setIntOnStore(store sdk.KVStore, key []byte, i int64) {
-	bz := make([]byte, 8)
-	n := binary.PutVarint(bz, i)
-	store.Set(key, bz[:n])
-}
-
-// check counter matches what's in store.
-// increment and store
-func incrementingCounter(t *testing.T, store sdk.KVStore, counterKey []byte, counter int64) (*sdk.Result, error) {
-	storedCounter := getIntFromStore(store, counterKey)
-	require.Equal(t, storedCounter, counter)
-	setIntOnStore(store, counterKey, counter+1)
-	return &sdk.Result{}, nil
-}
-
-// Number of messages doesn't matter to CheckTx.
-func TestMultiMsgCheckTx(t *testing.T) {
-	// TODO: ensure we get the same results
-	// with one message or many
-}
-
-// Interleave calls to Check and Deliver and ensure
-// that there is no cross-talk. Check sees results of the previous Check calls
-// and Deliver sees that of the previous Deliver calls, but they don't see eachother.
-func TestConcurrentCheckDeliver(t *testing.T) {
-	// TODO
-}
-
-func TestGRPCQuery(t *testing.T) {
-	grpcQueryOpt := func(bapp *BaseApp) {
-		testdata.RegisterQueryServer(
-			bapp.GRPCQueryRouter(),
-			testdata.QueryImpl{},
-		)
-	}
-
-	app := setupBaseApp(t, grpcQueryOpt)
-	app.GRPCQueryRouter().SetInterfaceRegistry(codectypes.NewInterfaceRegistry())
-
-	app.InitChain(abci.RequestInitChain{})
-	header := tmproto.Header{Height: app.LastBlockHeight() + 1}
-	app.BeginBlock(abci.RequestBeginBlock{Header: header})
-	app.Commit()
-
-	req := testdata.SayHelloRequest{Name: "foo"}
-	reqBz, err := req.Marshal()
-	require.NoError(t, err)
-
-	reqQuery := abci.RequestQuery{
-		Data: reqBz,
-		Path: "/testdata.Query/SayHello",
-	}
-
-	resQuery := app.Query(reqQuery)
-
-	require.Equal(t, abci.CodeTypeOK, resQuery.Code, resQuery)
-
-	var res testdata.SayHelloResponse
-	err = res.Unmarshal(resQuery.Value)
-	require.NoError(t, err)
-	require.Equal(t, "Hello foo!", res.Greeting)
-}
-
-// Test p2p filter queries
-func TestP2PQuery(t *testing.T) {
-	addrPeerFilterOpt := func(bapp *BaseApp) {
-		bapp.SetAddrPeerFilter(func(addrport string) abci.ResponseQuery {
-			require.Equal(t, "1.1.1.1:8000", addrport)
-			return abci.ResponseQuery{Code: uint32(3)}
-		})
-	}
-
-	idPeerFilterOpt := func(bapp *BaseApp) {
-		bapp.SetIDPeerFilter(func(id string) abci.ResponseQuery {
-			require.Equal(t, "testid", id)
-			return abci.ResponseQuery{Code: uint32(4)}
-		})
-	}
-
-	app := setupBaseApp(t, addrPeerFilterOpt, idPeerFilterOpt)
-
-	addrQuery := abci.RequestQuery{
-		Path: "/p2p/filter/addr/1.1.1.1:8000",
-	}
-	res := app.Query(addrQuery)
-	require.Equal(t, uint32(3), res.Code)
-
-	idQuery := abci.RequestQuery{
-		Path: "/p2p/filter/id/testid",
-	}
-	res = app.Query(idQuery)
-	require.Equal(t, uint32(4), res.Code)
-}
-
-func TestGetMaximumBlockGas(t *testing.T) {
-	app := setupBaseApp(t)
-	app.InitChain(abci.RequestInitChain{})
-	ctx := app.NewContext(true, tmproto.Header{})
-
-	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: 0}})
-	require.Equal(t, uint64(0), app.getMaximumBlockGas(ctx))
-
-	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: -1}})
-	require.Equal(t, uint64(0), app.getMaximumBlockGas(ctx))
-
-	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: 5000000}})
-	require.Equal(t, uint64(5000000), app.getMaximumBlockGas(ctx))
-
-	app.StoreConsensusParams(ctx, &abci.ConsensusParams{Block: &abci.BlockParams{MaxGas: -5000000}})
-	require.Panics(t, func() { app.getMaximumBlockGas(ctx) })
-}
-
-func TestListSnapshots(t *testing.T) {
-	setupConfig := &setupConfig{
-		blocks:             5,
-		blockTxs:           4,
-		snapshotInterval:   2,
-		snapshotKeepRecent: 2,
-		pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-	}
-
-	app, err := setupBaseAppWithSnapshots(t, setupConfig)
-	require.NoError(t, err)
-
-	resp := app.ListSnapshots(abci.RequestListSnapshots{})
-	for _, s := range resp.Snapshots {
-		assert.NotEmpty(t, s.Hash)
-		assert.NotEmpty(t, s.Metadata)
-		s.Hash = nil
-		s.Metadata = nil
-	}
-	assert.Equal(t, abci.ResponseListSnapshots{Snapshots: []*abci.Snapshot{
-		{Height: 4, Format: snapshottypes.CurrentFormat, Chunks: 2},
-		{Height: 2, Format: snapshottypes.CurrentFormat, Chunks: 1},
-	}}, resp)
-}
-
-func TestSnapshotWithPruning(t *testing.T) {
-	testcases := map[string]struct {
-		config            *setupConfig
-		expectedSnapshots []*abci.Snapshot
-		expectedErr       error
-	}{
-		"prune nothing with snapshot": {
-			config: &setupConfig{
-				blocks:             20,
-				blockTxs:           2,
-				snapshotInterval:   5,
-				snapshotKeepRecent: 1,
-				pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-			},
-			expectedSnapshots: []*abci.Snapshot{
-				{Height: 20, Format: 2, Chunks: 5},
-			},
-		},
-		"prune everything with snapshot": {
-			config: &setupConfig{
-				blocks:             20,
-				blockTxs:           2,
-				snapshotInterval:   5,
-				snapshotKeepRecent: 1,
-				pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningEverything),
-			},
-			expectedSnapshots: []*abci.Snapshot{
-				{Height: 20, Format: 2, Chunks: 5},
-			},
-		},
-		"default pruning with snapshot": {
-			config: &setupConfig{
-				blocks:             20,
-				blockTxs:           2,
-				snapshotInterval:   5,
-				snapshotKeepRecent: 1,
-				pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningDefault),
-			},
-			expectedSnapshots: []*abci.Snapshot{
-				{Height: 20, Format: 2, Chunks: 5},
-			},
-		},
-		"custom": {
-			config: &setupConfig{
-				blocks:             25,
-				blockTxs:           2,
-				snapshotInterval:   5,
-				snapshotKeepRecent: 2,
-				pruningOpts:        pruningtypes.NewCustomPruningOptions(12, 12),
-			},
-			expectedSnapshots: []*abci.Snapshot{
-				{Height: 25, Format: 2, Chunks: 6},
-				{Height: 20, Format: 2, Chunks: 5},
-			},
-		},
-		"no snapshots": {
-			config: &setupConfig{
-				blocks:           10,
-				blockTxs:         2,
-				snapshotInterval: 0, // 0 implies disable snapshots
-				pruningOpts:      pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-			},
-			expectedSnapshots: []*abci.Snapshot{},
-		},
-		"keep all snapshots": {
-			config: &setupConfig{
-				blocks:             10,
-				blockTxs:           2,
-				snapshotInterval:   3,
-				snapshotKeepRecent: 0, // 0 implies keep all snapshots
-				pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-			},
-			expectedSnapshots: []*abci.Snapshot{
-				{Height: 9, Format: 2, Chunks: 2},
-				{Height: 6, Format: 2, Chunks: 2},
-				{Height: 3, Format: 2, Chunks: 1},
-			},
-		},
-	}
-
-	for name, tc := range testcases {
-		t.Run(name, func(t *testing.T) {
-			app, err := setupBaseAppWithSnapshots(t, tc.config)
-
-			if tc.expectedErr != nil {
-				require.Error(t, err)
-				require.Equal(t, tc.expectedErr.Error(), err.Error())
-				return
-			}
-			require.NoError(t, err)
-
-			resp := app.ListSnapshots(abci.RequestListSnapshots{})
-			for _, s := range resp.Snapshots {
-				assert.NotEmpty(t, s.Hash)
-				assert.NotEmpty(t, s.Metadata)
-				s.Hash = nil
-				s.Metadata = nil
-			}
-			fmt.Println(resp)
-			assert.Equal(t, abci.ResponseListSnapshots{Snapshots: tc.expectedSnapshots}, resp)
-
-			// Validate that heights were pruned correctly by querying the state at the last height that should be present relative to latest
-			// and the first height that should be pruned.
-			//
-			// Exceptions:
-			//   * Prune nothing: should be able to query all heights (we only test first and latest)
-			//   * Prune default: should be able to query all heights (we only test first and latest)
-			//      * The reason for default behaving this way is that we only commit 20 heights but default has 100_000 keep-recent
-			var lastExistingHeight int64
-			if tc.config.pruningOpts.GetPruningStrategy() == pruningtypes.PruningNothing || tc.config.pruningOpts.GetPruningStrategy() == pruningtypes.PruningDefault {
-				lastExistingHeight = 1
-			} else {
-				// Integer division rounds down so by multiplying back we get the last height at which we pruned
-				lastExistingHeight = int64((tc.config.blocks/tc.config.pruningOpts.Interval)*tc.config.pruningOpts.Interval - tc.config.pruningOpts.KeepRecent)
-			}
-
-			// Query 1
-			res := app.Query(abci.RequestQuery{Path: fmt.Sprintf("/store/%s/key", capKey2.Name()), Data: []byte("0"), Height: lastExistingHeight})
-			require.NotNil(t, res, "height: %d", lastExistingHeight)
-			require.NotNil(t, res.Value, "height: %d", lastExistingHeight)
-
-			// Query 2
-			res = app.Query(abci.RequestQuery{Path: fmt.Sprintf("/store/%s/key", capKey2.Name()), Data: []byte("0"), Height: lastExistingHeight - 1})
-			require.NotNil(t, res, "height: %d", lastExistingHeight-1)
-			if tc.config.pruningOpts.GetPruningStrategy() == pruningtypes.PruningNothing || tc.config.pruningOpts.GetPruningStrategy() == pruningtypes.PruningDefault {
-				// With prune nothing or default, we query height 0 which translates to the latest height.
-				require.NotNil(t, res.Value, "height: %d", lastExistingHeight-1)
-			}
-		})
-	}
-}
-
-func TestLoadSnapshotChunk(t *testing.T) {
-	setupConfig := &setupConfig{
-		blocks:             2,
-		blockTxs:           5,
-		snapshotInterval:   2,
-		snapshotKeepRecent: 2,
-		pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-	}
-	app, err := setupBaseAppWithSnapshots(t, setupConfig)
-	require.NoError(t, err)
-
-	testcases := map[string]struct {
-		height      uint64
-		format      uint32
-		chunk       uint32
-		expectEmpty bool
-	}{
-		"Existing snapshot": {2, snapshottypes.CurrentFormat, 1, false},
-		"Missing height":    {100, snapshottypes.CurrentFormat, 1, true},
-		"Missing format":    {2, 3, 1, true},
-		"Missing chunk":     {2, snapshottypes.CurrentFormat, 9, true},
-		"Zero height":       {0, snapshottypes.CurrentFormat, 1, true},
-		"Zero format":       {2, 0, 1, true},
-		"Zero chunk":        {2, snapshottypes.CurrentFormat, 0, false},
-	}
-	for name, tc := range testcases {
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			resp := app.LoadSnapshotChunk(abci.RequestLoadSnapshotChunk{
-				Height: tc.height,
-				Format: tc.format,
-				Chunk:  tc.chunk,
-			})
-			if tc.expectEmpty {
-				assert.Equal(t, abci.ResponseLoadSnapshotChunk{}, resp)
-				return
-			}
-			assert.NotEmpty(t, resp.Chunk)
-		})
-	}
-}
-
-func TestOfferSnapshot_Errors(t *testing.T) {
-	// Set up app before test cases, since it's fairly expensive.
-	setupConfig := &setupConfig{
-		blocks:             0,
-		blockTxs:           0,
-		snapshotInterval:   2,
-		snapshotKeepRecent: 2,
-		pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-	}
-	app, err := setupBaseAppWithSnapshots(t, setupConfig)
-	require.NoError(t, err)
-
-	m := snapshottypes.Metadata{ChunkHashes: [][]byte{{1}, {2}, {3}}}
-	metadata, err := m.Marshal()
-	require.NoError(t, err)
-	hash := []byte{1, 2, 3}
-
-	testcases := map[string]struct {
-		snapshot *abci.Snapshot
-		result   abci.ResponseOfferSnapshot_Result
-	}{
-		"nil snapshot": {nil, abci.ResponseOfferSnapshot_REJECT},
-		"invalid format": {&abci.Snapshot{
-			Height: 1, Format: 9, Chunks: 3, Hash: hash, Metadata: metadata,
-		}, abci.ResponseOfferSnapshot_REJECT_FORMAT},
-		"incorrect chunk count": {&abci.Snapshot{
-			Height: 1, Format: snapshottypes.CurrentFormat, Chunks: 2, Hash: hash, Metadata: metadata,
-		}, abci.ResponseOfferSnapshot_REJECT},
-		"no chunks": {&abci.Snapshot{
-			Height: 1, Format: snapshottypes.CurrentFormat, Chunks: 0, Hash: hash, Metadata: metadata,
-		}, abci.ResponseOfferSnapshot_REJECT},
-		"invalid metadata serialization": {&abci.Snapshot{
-			Height: 1, Format: snapshottypes.CurrentFormat, Chunks: 0, Hash: hash, Metadata: []byte{3, 1, 4},
-		}, abci.ResponseOfferSnapshot_REJECT},
-	}
-	for name, tc := range testcases {
-		tc := tc
-		t.Run(name, func(t *testing.T) {
-			resp := app.OfferSnapshot(abci.RequestOfferSnapshot{Snapshot: tc.snapshot})
-			assert.Equal(t, tc.result, resp.Result)
-		})
-	}
-
-	// Offering a snapshot after one has been accepted should error
-	resp := app.OfferSnapshot(abci.RequestOfferSnapshot{Snapshot: &abci.Snapshot{
-		Height:   1,
-		Format:   snapshottypes.CurrentFormat,
-		Chunks:   3,
-		Hash:     []byte{1, 2, 3},
-		Metadata: metadata,
-	}})
-	require.Equal(t, abci.ResponseOfferSnapshot{Result: abci.ResponseOfferSnapshot_ACCEPT}, resp)
-
-	resp = app.OfferSnapshot(abci.RequestOfferSnapshot{Snapshot: &abci.Snapshot{
-		Height:   2,
-		Format:   snapshottypes.CurrentFormat,
-		Chunks:   3,
-		Hash:     []byte{1, 2, 3},
-		Metadata: metadata,
-	}})
-	require.Equal(t, abci.ResponseOfferSnapshot{Result: abci.ResponseOfferSnapshot_ABORT}, resp)
-}
-
-func TestApplySnapshotChunk(t *testing.T) {
-	setupConfig1 := &setupConfig{
-		blocks:             4,
-		blockTxs:           10,
-		snapshotInterval:   2,
-		snapshotKeepRecent: 2,
-		pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-	}
-	source, err := setupBaseAppWithSnapshots(t, setupConfig1)
-	require.NoError(t, err)
-
-	setupConfig2 := &setupConfig{
-		blocks:             0,
-		blockTxs:           0,
-		snapshotInterval:   2,
-		snapshotKeepRecent: 2,
-		pruningOpts:        pruningtypes.NewPruningOptions(pruningtypes.PruningNothing),
-	}
-	target, err := setupBaseAppWithSnapshots(t, setupConfig2)
-	require.NoError(t, err)
-
-	// Fetch latest snapshot to restore
-	respList := source.ListSnapshots(abci.RequestListSnapshots{})
-	require.NotEmpty(t, respList.Snapshots)
-	snapshot := respList.Snapshots[0]
-
-	// Make sure the snapshot has at least 3 chunks
-	require.GreaterOrEqual(t, snapshot.Chunks, uint32(3), "Not enough snapshot chunks")
-
-	// Begin a snapshot restoration in the target
-	respOffer := target.OfferSnapshot(abci.RequestOfferSnapshot{Snapshot: snapshot})
-	require.Equal(t, abci.ResponseOfferSnapshot{Result: abci.ResponseOfferSnapshot_ACCEPT}, respOffer)
-
-	// We should be able to pass an invalid chunk and get a verify failure, before reapplying it.
-	respApply := target.ApplySnapshotChunk(abci.RequestApplySnapshotChunk{
-		Index:  0,
-		Chunk:  []byte{9},
-		Sender: "sender",
-	})
-	require.Equal(t, abci.ResponseApplySnapshotChunk{
-		Result:        abci.ResponseApplySnapshotChunk_RETRY,
-		RefetchChunks: []uint32{0},
-		RejectSenders: []string{"sender"},
-	}, respApply)
-
-	// Fetch each chunk from the source and apply it to the target
-	for index := uint32(0); index < snapshot.Chunks; index++ {
-		respChunk := source.LoadSnapshotChunk(abci.RequestLoadSnapshotChunk{
-			Height: snapshot.Height,
-			Format: snapshot.Format,
-			Chunk:  index,
-		})
-		require.NotNil(t, respChunk.Chunk)
-		respApply := target.ApplySnapshotChunk(abci.RequestApplySnapshotChunk{
-			Index: index,
-			Chunk: respChunk.Chunk,
-		})
-		require.Equal(t, abci.ResponseApplySnapshotChunk{
-			Result: abci.ResponseApplySnapshotChunk_ACCEPT,
-		}, respApply)
-	}
-
-	// The target should now have the same hash as the source
-	assert.Equal(t, source.LastCommitID(), target.LastCommitID())
-}
-
-// func TestWithRouter(t *testing.T) {
-// 	// test increments in the ante
-// 	anteKey := []byte("ante-key")
-// 	anteOpt := func(bapp *BaseApp) { bapp.SetAnteHandler(anteHandlerTxTest(t, capKey1, anteKey)) }
-
-// 	// test increments in the handler
-// 	// deliverKey := []byte("deliver-key")
-// 	routerOpt := func(bapp *BaseApp) {
-// 		// bapp.SetRouter(&testCustomRouter{routes: sync.Map{}})
-// 		// r := sdk.NewRoute(routeMsgCounter, handlerMsgCounter(t, capKey1, deliverKey))
-// 		// bapp.Router().AddRoute(r)
-// 	}
-
-// 	app := setupBaseApp(t, anteOpt, routerOpt)
-// 	app.InitChain(abci.RequestInitChain{})
-
-// 	// Create same codec used in txDecoder
-// 	codec := codec.NewLegacyAmino()
-// 	registerTestCodec(codec)
-
-// 	nBlocks := 3
-// 	txPerHeight := 5
-
-// 	for blockN := 0; blockN < nBlocks; blockN++ {
-// 		header := tmproto.Header{Height: int64(blockN) + 1}
-// 		app.BeginBlock(abci.RequestBeginBlock{Header: header})
-
-// 		for i := 0; i < txPerHeight; i++ {
-// 			counter := int64(blockN*txPerHeight + i)
-// 			tx := newTxCounter(counter, counter)
-
-// 			txBytes, err := codec.Marshal(tx)
-// 			require.NoError(t, err)
-
-// 			res := app.DeliverTx(abci.RequestDeliverTx{Tx: txBytes})
-// 			require.True(t, res.IsOK(), fmt.Sprintf("%v", res))
-// 		}
-
-// 		app.EndBlock(abci.RequestEndBlock{})
-// 		app.Commit()
-// 	}
-// }
-
-func TestBaseApp_EndBlock(t *testing.T) {
-	db := dbm.NewMemDB()
-	name := t.Name()
-	logger := defaultLogger()
-
-	cp := &abci.ConsensusParams{
-		Block: &abci.BlockParams{
-			MaxGas: 5000000,
-		},
-	}
-
-	app := NewBaseApp(name, logger, db, nil)
-	app.SetParamStore(&paramStore{db: dbm.NewMemDB()})
-	app.InitChain(abci.RequestInitChain{
-		ConsensusParams: cp,
-	})
-
-	app.SetEndBlocker(func(ctx sdk.Context, req abci.RequestEndBlock) abci.ResponseEndBlock {
-		return abci.ResponseEndBlock{
-			ValidatorUpdates: []abci.ValidatorUpdate{
-				{Power: 100},
-			},
-		}
-	})
-	app.Seal()
-
-	res := app.EndBlock(abci.RequestEndBlock{})
-	require.Len(t, res.GetValidatorUpdates(), 1)
-	require.Equal(t, int64(100), res.GetValidatorUpdates()[0].Power)
-	require.Equal(t, cp.Block.MaxGas, res.ConsensusParamUpdates.Block.MaxGas)
 }
