+++ baseapp/testutil/messages.go
@@ -8,9 +8,15 @@ import (
 )
 
 func RegisterInterfaces(registry types.InterfaceRegistry) {
-	registry.RegisterImplementations((*sdk.Msg)(nil), &MsgCounter{}, &MsgCounter2{})
+	registry.RegisterImplementations(
+		(*sdk.Msg)(nil),
+		&MsgCounter{},
+		&MsgCounter2{},
+		&MsgKeyValue{},
+	)
 	msgservice.RegisterMsgServiceDesc(registry, &_Counter_serviceDesc)
 	msgservice.RegisterMsgServiceDesc(registry, &_Counter2_serviceDesc)
+	msgservice.RegisterMsgServiceDesc(registry, &_KeyValue_serviceDesc)
 }
 
 var _ sdk.Msg = &MsgCounter{}
@@ -31,4 +37,17 @@ func (msg *MsgCounter2) ValidateBasic() error {
 		return nil
 	}
 	return sdkerrors.Wrap(sdkerrors.ErrInvalidSequence, "counter should be a non-negative integer")
+}
+
+var _ sdk.Msg = &MsgKeyValue{}
+
+func (msg *MsgKeyValue) GetSigners() []sdk.AccAddress { return []sdk.AccAddress{} }
+func (msg *MsgKeyValue) ValidateBasic() error {
+	if msg.Key == nil {
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "key cannot be nil")
+	}
+	if msg.Value == nil {
+		return sdkerrors.Wrap(sdkerrors.ErrInvalidRequest, "value cannot be nil")
+	}
+	return nil
 }
