+++ baseapp/testutil/messages.proto
@@ -12,16 +12,27 @@ message MsgCounter {
 }
 
 message MsgCounter2 {
-  int64 counter       = 1;
+  int64 counter         = 1;
   bool  fail_on_handler = 2;
 }
 
+message MsgCreateCounterResponse {}
+
+message MsgKeyValue {
+  bytes key   = 1;
+  bytes value = 2;
+}
+
+message MsgCreateKeyValueResponse {}
+
 service Counter {
-    rpc IncrementCounter(MsgCounter) returns (MsgCreateCounterResponse);
+  rpc IncrementCounter(MsgCounter) returns (MsgCreateCounterResponse);
 }
 
 service Counter2 {
-    rpc IncrementCounter(MsgCounter2) returns (MsgCreateCounterResponse);
+  rpc IncrementCounter(MsgCounter2) returns (MsgCreateCounterResponse);
 }
 
-message MsgCreateCounterResponse {}
+service KeyValue {
+  rpc Set(MsgKeyValue) returns (MsgCreateKeyValueResponse);
+}
